Make json_tree actually work,
But I haven't made up my mind how to fix printing a tree.

As of now the tree can be built by adding data to a tokenizer and if it
fails, add some more data and restart from the node returned in the
error.

Also, it should be possible to add data in the tokenizer release
callback, and hence allways just running over the tree once.