Add external handling for function containers
This sheds some interesting light on the new approach. Since it is now
utilizing parial template specialisation, then looking up superclasses
does not work. Ie if you inherit an object with JT_* meta information,
then this will not be available unless you reexpose it in the inherited
class. Old behavior still holds if meta information is declared inline.

This is true for both JT_STRUCT* and JT_FUNCTION* families.