Remove unused function warnings
GCC docs states for -Wunused-function: Warn whenever a static function is
declared but not defined or a non-inline static function is unused.
Marking the functions as inline will therefore remove the warnings. This
works in Apple Clang in Xcode 11 as well and is likely to work in
regular Clang also.